+++ lib/database/src/sb/projectors/SbCylinderSectionProjector.c++
@@ -319,7 +319,7 @@ SbCylinderSectionProjector::getRotation(const SbVec3f &p1, const SbVec3f &p2) {
     if ((cosAngle > 1.0) || (cosAngle < -1.0))
         return SbRotation::identity();
 
-    float angle = acosf(cosAngle);
+    float angle = std::acos(cosAngle);
 
     // This will either be the same as the cylinder's
     // axis, or the same but with direction reversed
