+++ lib/database/src/so/elements/SoGLTextureMatrixElement.c++
@@ -174,7 +174,7 @@ SoGLTextureMatrixElement::rotateEltBy(const SbRotation &rotation)
     rotation.getValue(axis, angle);
 
     glMatrixMode(GL_TEXTURE);
-    glRotatef(angle * (180.0 / M_PI), axis[0], axis[1], axis[2]);
+    glRotatef(angle * float(180.0 / M_PI), axis[0], axis[1], axis[2]);
     glMatrixMode(GL_MODELVIEW);
 
     SoTextureMatrixElement::rotateEltBy(rotation);
